extend typeMapping, add JSON data for more icons
The type mapping for icons now supports a translatable title text. This
is put in the title attribute of the icon node and should appear at
tooltip on hover.
Instead of pure text it is now possible to provide a JSON array of
config objects as display data. Each entry has either a 'type' property
(refering to the typeMapping, thus showing an icon) or a 'text' property
(just show this text).